Add About page and static pages specifications
Without this patch there is no About page and we don't have any
specification for how the  static page should be laid out. This is a
problem because we cannot make changes to the site without without a way
to automatically verify that the site behaves as expected. We added some
examples and we're using the rails layout to render each static page
without duplication

This patch completes up to and including RailsTutorial section 3.3.4

This patch includes pry for debugging spec failures using:

    require 'pry'; binding.pry